more comments, removed ModelCopy, placed the copy in the JCM, made modelCopyTest concrete
